refactor(tests): move fixtures to their scope-local conftest files
- Move tmp_repository from root conftest to 01_loader/conftest.py
- Move tmp_class_repository from root conftest to 02_models/conftest.py
- Extract copy_and_update_repository and update_tmp_repository helpers
  to tests/helpers.py to make them importable across conftest files
- Remove shutil import from root conftest (no longer used there)
a Singleton (patterns.Singleton). The loader tests initialize the Singleton
with the correct application context (fsm_tools). Config tests inherit
this instance. Reversing the order causes 54 test failures.

Refs: #32, #41
